fix(database,config): misunderstood how transport encrypt works for mssql dsn
(refer to comment id its-go/database/service.go-1)

in the making of this encrypt configuration that works between mssql and postgresql, we decided that on sqlserver the usage definiton of "disabled" and "false" is confusing. as above, "disable" is defined as "Data send between client and server is not encrypted", while "false" is "Data sent between client and server is not encrypted [[beyond the login packet]].". the consequences of using "false" instead of "disable" is you may get tls cert verify error like below:

```
[error] failed to initialize database, got error TLS Handshake failed: tls: failed to verify certificate: x509: cannot validate certificate for (db ip/host) because it doesn't contain any IP SANs
```

"[[beyond the login packet]]" implies there's transport encryption that is enabled at login. if you specify TransportEncrypt as either "false" or "login-only", and you get tls handshake verify certificate error, make sure to configure TrustServerCertificate as "true".